Refactor code duplication from package wrappers
