+++ src/package/PackageWithFreezableDirectories.js
@@ -4,6 +4,7 @@ import Contracts from '../utils/Contracts'
 import Package from './Package'
 import PackageDeployer from './PackageDeployer'
 import PackageProvider from './PackageProvider'
+import ImplementationDirectoryDeployer from '../directory/ImplementationDirectoryDeployer'
 
 const log = new Logger('Package')
 
@@ -11,7 +12,7 @@ export default class PackageWithFreezableDirectories extends Package {
 
   static async fetch(address, txParams = {}) {
     const provider = new PackageProvider(txParams)
-    return await provider.fetchForFrezzableDirectories(address)
+    return await provider.fetchForFreezableDirectories(address)
   }
 
   static async deploy(txParams = {}) {
@@ -19,20 +20,13 @@ export default class PackageWithFreezableDirectories extends Package {
     return await deployer.deployForFreezableDirectories()
   }
 
-  async getImplementationDirectory(version) {
-    const directoryAddress = await this.package.getVersion(version)
-    const FreezableImplementationDirectory = Contracts.getFromLib('FreezableImplementationDirectory')
+  async wrapImplementationDirectory(directoryAddress) {
+    const FreezableImplementationDirectory = Contracts.getFromLib('FreezableImplementationDirectory');
     return new FreezableImplementationDirectory(directoryAddress)
   }
 
-  async newVersion(version) {
-    log.info('Adding new version...')
-    const FreezableImplementationDirectory = Contracts.getFromLib('FreezableImplementationDirectory')
-    const directory = await FreezableImplementationDirectory.new(this.txParams)
-    log.info(`App directory created at ${directory.address}`)
-    await this.package.addVersion(version, directory.address, this.txParams)
-    log.info(`Added version ${version}`)
-    return directory
+  async newDirectory() {
+    return ImplementationDirectoryDeployer.freezable(this.txParams).deployLocal()
   }
 
   async isFrozen(version) {
