+++ src/package/PackageWithNonFreezableDirectories.js
@@ -1,11 +1,9 @@
-import Logger from '../utils/Logger'
 import Contracts from '../utils/Contracts'
 
 import Package from './Package'
 import PackageDeployer from './PackageDeployer'
 import PackageProvider from './PackageProvider'
-
-const log = new Logger('Package')
+import ImplementationDirectoryDeployer from "../directory/ImplementationDirectoryDeployer";
 
 export default class PackageWithNonFreezableDirectories extends Package {
 
@@ -19,19 +17,12 @@ export default class PackageWithNonFreezableDirectories extends Package {
     return await deployer.deploy()
   }
 
-  async getImplementationDirectory(version) {
-    const directoryAddress = await this.package.getVersion(version)
-    const ImplementationDirectory = Contracts.getFromLib('ImplementationDirectory')
+  async wrapImplementationDirectory(directoryAddress) {
+    const ImplementationDirectory = Contracts.getFromLib('ImplementationDirectory');
     return new ImplementationDirectory(directoryAddress)
   }
 
-  async newVersion(version) {
-    log.info('Adding new version...')
-    const ImplementationDirectory = Contracts.getFromLib('ImplementationDirectory')
-    const directory = await ImplementationDirectory.new(this.txParams)
-    log.info(`App directory created at ${directory.address}`)
-    await this.package.addVersion(version, directory.address, this.txParams)
-    log.info(`Added version ${version}`)
-    return directory
+  async newDirectory() {
+    return ImplementationDirectoryDeployer.nonFreezable(this.txParams).deployLocal()
   }
 }
